Moved graph tree up a level
