+++ src/layout/layout_tree/mod.rs
@@ -1,4 +1,3 @@
 pub mod layout;
 pub mod container;
 pub mod tree;
-pub mod graph_tree;
